Árak egész számban jelenjenek meg. Két sor átírásával megoldottam, hogy az állatok ára egész számként jelenjen meg: modellben a 'Price' legyen int, service-ben pedig castolja a decimalt.

+++ AnimalView_Modul/AnimalView_Modul/Services/AnimalService.cs
@@ -113,7 +113,7 @@ namespace AnimalView.Dnn.AnimalView_Modul.Services
                 AnimalId = RawAnimal.Bvin,
                 Name = AName,
                 Image = RawAnimal.ImageFileMedium,
-                Price = RawAnimal.SitePrice,
+                Price = (int) RawAnimal.SitePrice,
                 BirthDate = ABirthDate,
                 Gender = AGender,
                 Genetics = AGenetics,
